+++ packages/callout/src/js/react.js
@@ -7,7 +7,7 @@
  *
  **************************************************************************************************************************************************************/
 
-import React, { Fragment } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 
@@ -20,38 +20,56 @@ import PropTypes from 'prop-types';
 /**
  * Default callout
  *
+ * @param  {string}  title            - The title of the header
+ * @param  {string}  level            - The tag level (<h1/> <h2/> etc), default: '1'
+ * @param  {boolean} srOnlyTitle      - Title is visible to screen readers only
  * @param  {boolean} dark             - Add the dark variation class, optional
  * @param  {boolean} alt              - Add the alt variation class, optional
- * @param  {string}  description      - A description of the content of the callout for a11y
  * @param  {string}  children         - Anything inside
  * @param  {string}  className        - An additional class, optional
  * @param  {object}  attributeOptions - Any other attribute options
  */
-export const AUcallout = ({ dark, alt, description, children, className = '', ...attributeOptions }) => (
-	<section
-		className={ `au-callout ${ className }${ dark ? ' au-callout--dark' : '' }${ alt ? ' au-callout--alt' : '' }` }
-		aria-label={ description }
-		{ ...attributeOptions }
-	>
-		{ children }
-	</section>
-);
+export const AUcallout = ({ title, level, srOnlyTitle, dark, alt, description, children, className = '', ...attributeOptions }) => {
+	const HeadingTag = `h${ level }`;
+
+	return (
+		<section
+			className={ `au-callout ${ className }${ dark ? ' au-callout--dark' : '' }${ alt ? ' au-callout--alt' : '' }` }
+			{ ...attributeOptions }
+		>
+			<HeadingTag
+				children={ title }
+				className={ `au-callout__heading${ srOnlyTitle ? ' au-callout__heading--sronly' : '' }` }
+			/>
+			{ children }
+		</section>
+	)
+};
 
 AUcallout.propTypes = {
+	title: PropTypes.string.isRequired,
+	level: PropTypes.number,
+	srOnlyTitle: PropTypes.bool,
 	dark: PropTypes.bool,
 	alt: PropTypes.bool,
-	description: PropTypes.string.isRequired,
 	children: PropTypes.node.isRequired,
 	className: PropTypes.string,
+};
+
+AUcallout.defaultProps = {
+	level: 2,
+	srOnlyTitle: false,
 };
 
 
 /**
  * Calendar callout
  *
+ * @param  {string}  title            - The title of the header
+ * @param  {string}  level            - The tag level (<h1/> <h2/> etc), default: '1'
+ * @param  {boolean} srOnlyTitle      - Title is visible to screen readers only
  * @param  {boolean} dark             - Add the dark variation class, optional
  * @param  {boolean} alt              - Add the alt variation class, optional
- * @param  {string}  description      - A description of the content of the callout for a11y
  * @param  {string}  subline          - The subline of the event, optional
  * @param  {string}  datetime         - The datetime of the event as ISO datetime
  * @param  {string}  time             - The time that appears on the page
@@ -59,25 +77,39 @@ AUcallout.propTypes = {
  * @param  {string}  className        - An additional class, optional
  * @param  {object}  attributeOptions - Any other attribute options
  */
-export const AUcalloutCalendar = ({ dark, alt, description, subline, datetime, time, name, className = '', ...attributeOptions }) => (
-	<section
-		className={ `au-callout au-callout--calendar-event ${ className }${ dark ? ' au-callout--dark' : '' }${ alt ? ' au-callout--alt' : '' }` }
-		aria-label={ description }
-		{ ...attributeOptions }
-	>
-		{ subline && <span className="au-callout--calendar-event__lede">{ subline }</span> }
-		<time className="au-callout--calendar-event__time" dateTime={ new Date( datetime ).toJSON() }>{ time }</time>
-		{ name && <span className="au-callout--calendar-event__name">{ name }</span> }
-	</section>
-);
+export const AUcalloutCalendar = ({ title, level, srOnlyTitle, dark, alt, subline, datetime, time, name, className = '', ...attributeOptions }) => {
+	const HeadingTag = `h${ level }`;
+
+	return (
+		<section
+			className={ `au-callout au-callout--calendar-event ${ className }${ dark ? ' au-callout--dark' : '' }${ alt ? ' au-callout--alt' : '' }` }
+			{ ...attributeOptions }
+		>
+			<HeadingTag
+				children={ title }
+				className={ `au-callout__heading${ srOnlyTitle ? ' au-callout__heading--sronly' : '' }` }
+			/>
+			{ subline && <p className="au-callout--calendar-event__lede">{ subline }</p> }
+			<time className="au-callout--calendar-event__time" dateTime={ new Date( datetime ).toJSON() }>{ time }</time>
+			{ name && <span className="au-callout--calendar-event__name">{ name }</span> }
+		</section>
+	)
+}
 
 AUcalloutCalendar.propTypes = {
+	title: PropTypes.string.isRequired,
+	level: PropTypes.number,
+	srOnlyTitle: PropTypes.bool,
 	dark: PropTypes.bool,
 	alt: PropTypes.bool,
-	description: PropTypes.string.isRequired,
 	subline: PropTypes.string,
 	datetime: PropTypes.string.isRequired,
 	time: PropTypes.string.isRequired,
 	name: PropTypes.string,
 	className: PropTypes.string,
+};
+
+AUcalloutCalendar.defaultProps = {
+	level: 2,
+	srOnlyTitle: true,
 };