+++ packages/callout/CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ## Versions
 
+* [v3.0.0 - Replace aria-label with a title](#v300)
 * [v2.1.0 - Remove `margin-bottom`, increase `margin-top` for paragraphs following callout](#v210)
 * [v2.0.8 - Update dependencies](#v208)
 * [v2.0.7 - Removing web pack dev server, updating dependencies](#v207)
@@ -36,6 +37,11 @@
 
 
 ## Release History
+
+### v3.0.0
+
+- Replace aria-label with a title
+
 
 ### v2.1.0